:warning: Remove mean motion time-derivative from J4 props
The Kozai's theory used to code the J4 orbit propagators was not
developed to consider a time-derivative of the mean motion. Vallado's
uses this factors in a simplified example. In our case, the assumptions
will break with a very small propagation period. Hence, we remove to
make the theory consistent.